requirements.txt
discord~=2.1.0
python-dotenv~=0.21.0
